Moved compatibility assert_allclose to astropy.tests.compat to avoid importing Numpy when running egg_info.
