+++ astropy/units/tests/test_format.py
@@ -5,7 +5,8 @@ Regression tests for the units.format package
 
 from __future__ import absolute_import, unicode_literals, division, print_function
 
-from astropy.tests.helper import raises, assert_allclose
+from ...tests.helper import raises
+from ...tests.compat import assert_allclose
 
 from ... import units as u
 from .. import core
